requirements.txt
google-generativeai==0.8.3
python-decouple==3.8